requirements.txt
Django==1.8.5
Genshi==0.7
Markups==0.6.3
PyJWT==1.4.0
Pygments==2.0.2
argparse==1.2.1
certifi==2015.9.6.2
cssselect==0.9.1
django-messages==0.5.1
docutils==0.12
django-haystack==2.1.0
elasticsearch==1.8.0
gunicorn==19.3.0
lxml==3.4.4
oauthlib==1.0.3
pefile==1.2.10.post114
pyquery==1.2.9
python-openid==2.2.5
python-ptrace==0.8.1
python-social-auth==0.2.13
requests==2.8.1
requests-oauthlib==0.5.0
rst2html5==1.7.5
sh==1.11
simplejson==3.8.0
six==1.10.0
urllib3==1.12
waliki==0.5
wsgiref==0.1.2
